Update Media.tsx
Updated the Media Component to detect if the Entry Picture is a URL or a Path to an image.

This was causing an issue where the Google Home playing music would show a cover image would point to a YouTube thumbnail URL. Before this update it would prepend the Hassio Address, Causing it to fail. Now it will check to see if it has HTTP in it, if it does then it is a URL address, otherwise, it would assume its a file path.